bsp: fs: change to return the read size
The read size is required if we read a binary data instead of string data.
This is needed for that we are moving the encryption module out of the read function
because cipher data is binary.

Also the parameter type is changed from 'unsigned int' to 'size_t' for read and write.

Signed-off-by: Sanghee Kim <sh0130.kim@samsung.com>