security: move sw backed fs encryption into bsp
Currently there is a problem that fs initial encryption is failed
at first booting. It doesn't encrypted.

And it's hard to handle I/O operation in security storage system because
raw flash operation and partition configuration is very different
in each chipset.

Therefore it's good to implement sw encryption integration in bsp.

This patch includes revert of following commit.
8f4324d bsp: esp8266: changed to use sw encryption library in security system